requirements.txt
# Agent 项目依赖
# Python 3.9+ 要求

# ==================== Web 框架 ====================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # 文件上传支持

# ==================== LangChain 生态 ====================
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.40
langchain-openai>=0.0.5  # 同时支持 DeepSeek 和 OpenAI 兼容接口

# ==================== 数据库 ====================
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0  # pymysql 需要

# ==================== 数据验证 ====================
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ==================== 向量数据库（可选使用） ====================
# chromadb>=0.4.0  # 暂时注释掉，等7阶段再添加

# ==================== 配置工具 ====================
pyyaml>=6.0
python-dotenv>=1.0.0

# ==================== 文档解析库（可选使用） ====================
# pypdf2>=3.0.0
# python-docx>=0.8.11
# openpyxl>=3.1.0

# ==================== Web 自动化（可选使用） ====================
# playwright>=1.40.0

# ==================== HTTP 客户端 ====================
httpx>=0.25.0
aiohttp>=3.9.0

# ==================== 日志增强 ====================
colorlog>=6.7.0
python-json-logger>=2.0.7

# ==================== 工具库 ====================
tenacity>=8.2.3  # 重试库

# ==================== 音频处理（可选） ====================
# 如果需要本地语音识别，可以安装以下库：
# SpeechRecognition>=3.10.0
# pydub>=0.25.1
# pyaudio>=0.2.11

